pyproject.toml
[project]
name = "r-ced"
version = "0.1.0"
description = "Faithful Implementation of R-CED Network"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.1",
    "numpy>=2.2.5",
    "omegaconf>=2.3.0",
    "pandas>=2.2.3",
    "pesq>=0.0.4",
    "pystoi>=0.4.1",
    "soundfile>=0.13.1",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.7",
]
